Getting modules from database and adding modules from frontend (#52)
* Refactor, Redux, and Redux-Sagas

Variables such as Modules are available on Redux. Redux-Sagas is not being called yet. I need to see what I am missing.

* Minor Fixes

Everything from previous PR is now in Redux-Sagas. There is an error in the console regarding the dialog but it seems it will get fixed if we upgrade to mui

* Upgrade to mui

MUI imports and some linting changes

* MUI Migration complete

Everything looks the same to me.

* Minor Changes to adding a Module

Now there is a textfield for module number

* Updated routes for requests

* Getting modules from the database

* Refactoring

* More refactoring

* Module Dialog supports Adding and Editing

I generalized ModuleDialog display different messages depending if we are intending to add or edit a module. But editing is not yet enabled

* Messages display after action

Used a snackbar to display the latest action's result

* Action feedback improved

Snackbar is smarter about the message, the type of alert, and when to display it.

* Merge complete, adding a module works

Users are not being logged out yet

* Made auth token for requests persistent

The method to set the token in axios only worked until the page was closed or refreshed. With this, now I can come back later and it still works. But this solution is not final.

* Added testing for module state

I finished the things I had left to do for this PR, now all the console.logs should be gone, please tell me if there are any remaining.